some sdk changes and additions, minor code changes, added libraries: 'miniz', 'CURL', 'OpenSSL' and their licenses. updated some variables using pointers(* ) to std::unique_ptr. added more pcb.h error types and more debugging variables inside CMakeLists.txt. updated Framebuffer creation, added more settings. more efficient memory management inside WorldRenderer.cpp. sleep well luna!

+++ src/me/sleepyfish/mcclone/entity/player/EnumPlayerModelParts.h
@@ -23,14 +23,14 @@ enum class EnumPlayerModelParts {
 
 };
 
+/**
+ * @author SleepyFish
+ */
 struct PlayerModelPartInfo {
 
-    int partId;
-
+    uint8_t partId;
     int partMask;
-
     std::string partName;
-
     std::string translationKey;
 
 };
